A simple UITableView, with add, edit, and delete.  No moving, although, that's not too hard.
Separate classes for the UITableViewDelegate and UITableViewDataSource, which is
a little more fine grained than I normally do it, but for explanation, it might
be better.

I'm not super happy with the encapsulation.  The data source has the table view
as a member.  It's because there needs to be a reloadData() call on the table
when things are changed, and the data source knows when a reload is necessary.